+++ src/device_list.cpp
@@ -15,7 +15,7 @@
 #include "sdk/android/native_api/jni/jvm.h"
 #endif
 
-#if defined(SORA_UNITY_SDK_MACOS) || defined(SORA_UNITY_SDK_IOS)
+#if defined(SORA_UNITY_SDK_MACOS) || defined(SORA_UNITY_SDK_IOS) || !defined(SORA_UNITY_SDK_VISIONOS)
 #include <sora/mac/mac_capturer.h>
 #elif defined(SORA_UNITY_SDK_ANDROID)
 #include <sora/android/android_capturer.h>
@@ -26,7 +26,7 @@ namespace sora_unity_sdk {
 
 bool DeviceList::EnumVideoCapturer(
     std::function<void(std::string, std::string)> f) {
-#if defined(SORA_UNITY_SDK_MACOS) || defined(SORA_UNITY_SDK_IOS)
+#if defined(SORA_UNITY_SDK_MACOS) || defined(SORA_UNITY_SDK_IOS) || !defined(SORA_UNITY_SDK_VISIONOS)
 
   return sora::MacCapturer::EnumVideoDevice(f);
 
@@ -67,8 +67,8 @@ bool DeviceList::EnumVideoCapturer(
 bool DeviceList::EnumAudioRecording(
     std::function<void(std::string, std::string)> f) {
   auto env = webrtc::CreateEnvironment();
-#if defined(SORA_UNITY_SDK_ANDROID) || defined(SORA_UNITY_SDK_IOS)
-  // Android や iOS の場合常に１個しかなく、かつ adm->RecordingDeviceName() を呼ぶと fatal error が起きるので
+#if defined(SORA_UNITY_SDK_ANDROID) || defined(SORA_UNITY_SDK_IOS) || defined(SORA_UNITY_SDK_VISIONOS)
+  // Android や iOS、visionOS の場合常に１個しかなく、かつ adm->RecordingDeviceName() を呼ぶと fatal error が起きるので
   // 適当な名前で１回だけコールバックする
   f("0", "0");
   return true;
@@ -119,8 +119,8 @@ bool DeviceList::EnumAudioRecording(
 bool DeviceList::EnumAudioPlayout(
     std::function<void(std::string, std::string)> f) {
   auto env = webrtc::CreateEnvironment();
-#if defined(SORA_UNITY_SDK_ANDROID) || defined(SORA_UNITY_SDK_IOS)
-  // Android や iOS の場合常に１個しかなく、かつ adm->PlayoutDeviceName() を呼ぶと fatal error が起きるので
+#if defined(SORA_UNITY_SDK_ANDROID) || defined(SORA_UNITY_SDK_IOS) || defined(SORA_UNITY_SDK_VISIONOS)
+  // Android や iOS、visionOS の場合常に１個しかなく、かつ adm->PlayoutDeviceName() を呼ぶと fatal error が起きるので
   // 適当な名前で１回だけコールバックする
   f("0", "0");
   return true;